refactor(capabilities): :recycle: migrate easter-eggs to component-store pattern
Flatten features/easter-eggs/ components/ + hooks/ into per-component
folders each with an index.ts barrel and a use-*-store.ts hook.

- center-container: pure/presentational, no store
- white-rabbit: pure utility fn, no store
- dejavu: stateful; useDejavu → useDejavuStore returning ComponentStore
- neo-room-easter-egg: stateful; inline useState extracted to useNeoRoomEasterEggStore

Update importers in brand-header.tsx and command-palette.tsx to use
the new barrel paths.
