+++ src/components/design-system/organism/command-palette/command-palette.tsx
@@ -9,7 +9,7 @@ import { motion } from "framer-motion";
 import { trackWith } from "@/lib/tracking/tracking";
 import { tracking } from "@/types/configuration/tracking";
 import { slugs } from "@/types/configuration/slug";
-import { whiteRabbitEasterEgg } from "@/components/features/easter-eggs/components/white-rabbit";
+import { whiteRabbitEasterEgg } from "@/components/features/easter-eggs/white-rabbit";
 import { Command } from "cmdk";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/navigation";
@@ -27,7 +27,7 @@ import { CustomizeMatrixRainItem } from "./customize-matrix-rain-item";
 
 const NeoRoomEasterEgg = dynamic(
   () =>
-    import("@/components/features/easter-eggs/components/neo-room-easter-egg"),
+    import("@/components/features/easter-eggs/neo-room-easter-egg"),
   { ssr: false },
 );
