+++ src/components/design-system/organism/header/brand-header.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { MatrixHeaderBackground } from "@/components/design-system/molecules/effects/matrix-header-background";
-import { DejavuEasterEgg } from "@/components/features/easter-eggs/components/dejavu";
+import { DejavuEasterEgg } from "@/components/features/easter-eggs/dejavu";
 import { FC } from "react";
 import logoImage from "../../../../../public/media/logo.png";
 import { ImageGlow } from "../../atoms/effects/image-glow";
