+++ src/components/features/easter-eggs/center-container/center-container.tsx
