+++ src/components/features/easter-eggs/center-container/index.ts
@@ -1,0 +1,1 @@
+export { CenterContainer } from "./center-container";
