+++ src/components/features/easter-eggs/dejavu/index.ts
@@ -1,0 +1,1 @@
+export { DejavuEasterEgg } from "./dejavu";
