+++ src/components/features/easter-eggs/neo-room-easter-egg/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./neo-room-easter-egg";
