+++ src/components/features/easter-eggs/white-rabbit/index.ts
@@ -1,0 +1,1 @@
+export { whiteRabbitEasterEgg } from "./white-rabbit";
